requirements.txt
flask>=3.0.0
anthropic>=0.30.0
firebase-admin>=6.5.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

#pdfplumber>=0.10.0
#//langchain-core>=0.1.0
#//langchain-huggingface>=0.0.3
#//langchain-pinecone>=0.1.0
#//pinecone-client>=3.0.0